refactor(website): move tailwind config to workspace root
This will allow us to share it as a preset across multiple projects.

ci(website): disable preflight only for project
